Conflict analysis
Changes to how a T-inconsistency is handled in the DPLL engine. It's still not
working correctly, though. This commit fixes a problem, where the UIP analysis
didn't work properly for theory conflicts.
E.g.:
x @ 2 --> a = b @ 2
                    \
                     --> Conflict
                    /
-y @ 1 --> a != b @ 1

With a propositional conflict, different decision levels right before the
conflict cannot happen.
The conflict clause/learnt clause is computed as the negated conjunction of the
theory explanation and adding the literal causing the conflict. Finally all but
one literal per decision level are removed.