Proof code rewritten to Int + code reorganization
Got rid of the ProofStructureNode class and changed the code accordingly. Now
the parents are stored in proofStructure and the edge labels in proofLabels.
Both are Array[Int].

Added a comment to the TheorySolver interface, stating that adheres to the one
in the DPLL(T) paper.

Reorganized the CongruenceClosure class, such that all vars/vals are at the top,
followed by the classes in approximate call order.